+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [1.5.0] - Unreleased
 
 * Bugs
+    * Data comparison of two sections would say the data differs even when there are 0 different rows or columns.
     * Removed 1 pixel of unwanted gap between the top of the tabular results and the top border. It made it possible to see the text that was being scrolled, super annoying.
 
 ## [1.4.1] - 2026-02-03